- Updated to Angelscript 2.7.0 (this has added support for 64bit cpu's) - Fixed lots of GCC warnings.
git-svn-id: https://fuji.googlecode.com/svn/trunk@336 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Middleware/angelscript/source/as_scriptnode.cpp
@@ -72,7 +72,7 @@ void asCScriptNode::SetToken(sToken *token)
 	tokenType   = token->type;
 }
 
-void asCScriptNode::UpdateSourcePos(int pos, int length)
+void asCScriptNode::UpdateSourcePos(size_t pos, size_t length)
 {
 	if( pos == 0 && length == 0 ) return;
